services/ai-orchestrator/requirements.txt
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
poml>=1.0.0
aiofiles>=23.0.0